backend/requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
pandas
openpyxl
djangorestframework-simplejwt